+++ CHANGELOG.md
@@ -11,6 +11,8 @@
   [#914](https://github.com/owncloud/music/issues/914)
 - Possibility to start playback and/or set shuffle/repeat with the URL arguments
   [#922](https://github.com/owncloud/music/issues/922)
+- Option to remove duplicates from a playlist
+  [#690](https://github.com/owncloud/music/issues/690)
 
 ### Changed
 - Allow replacing '/' and characters forbidden on Windows file names with '_' when matching image files to artist names